--- frontend/packages/process-editor/src/ProcessEditor.module.css
@@ -1,13 +1,0 @@
-.alertWrapper {
-  display: flex;
-  justify-content: center;
-  margin-block: 1rem;
-}
-
-.alert {
-  width: fit-content;
-}
-
-.alertText {
-  white-space: pre-line;
-}
