+++ frontend/packages/process-editor/src/ProcessEditor.tsx
@@ -1,11 +1,9 @@
 import React from 'react';
-import classes from './ProcessEditor.module.css';
-import { Trans, useTranslation } from 'react-i18next';
-import { Alert, Heading, Link, Paragraph } from '@digdir/design-system-react';
+import { useTranslation } from 'react-i18next';
+import { Alert, Heading, Paragraph } from '@digdir/design-system-react';
 import { PageLoading } from './components/PageLoading';
 import { Canvas } from './components/Canvas';
 import { BpmnContextProvider } from './contexts/BpmnContext';
-import { getIfVersionIs8OrNewer } from './utils/processEditorUtils';
 
 export type ProcessEditorProps = {
   bpmnXml: string | undefined | null;
@@ -28,38 +26,9 @@ export const ProcessEditor = ({
     return <NoBpmnFoundAlert />;
   }
 
-  const isEditAllowed: boolean = getIfVersionIs8OrNewer(appLibVersion);
-
   return (
     <BpmnContextProvider bpmnXml={bpmnXml}>
-      {!isEditAllowed && (
-        <div className={classes.alertWrapper}>
-          {/* TODO - Add logic check for edit button*/}
-          <Alert severity='warning' className={classes.alert}>
-            <Heading level={1} size='xsmall' spacing>
-              {t('process_editor.too_old_version_title')}
-            </Heading>
-            <Paragraph className={classes.alertText} size='small'>
-              <Trans i18nKey={t('process_editor.too_old_version_text')}>
-                Applikasjonen din har versjon <strong>{appLibVersion}</strong> av app-lib-pakken.
-                <br />
-                Versjonen er for gammel for å kunne redigere denne modellen. <br />
-                <br />
-                Du kan se på prosessen, men vi anbefaler at du oppgraderer til versjon 8 eller
-                nyere.
-                <br />
-                <br />
-                Trenger du hjelp til å oppgradere, kan du kontakte{' '}
-                <Link href='servicedesk@altinn.no' target='_new' rel='noopener noreferrer'>
-                  servicedesken
-                </Link>{' '}
-                vår.
-              </Trans>
-            </Paragraph>
-          </Alert>
-        </div>
-      )}
-      <Canvas onSave={onSave} isEditAllowed={isEditAllowed} />
+      <Canvas onSave={onSave} appLibVersion={appLibVersion} />
     </BpmnContextProvider>
   );
 };
