+++ frontend/packages/process-editor/src/components/Canvas/Canvas.module.css
@@ -2,13 +2,22 @@
   --mask-icon-size: 0 0 /46px 46px;
 }
 
+.wrapper {
+  display: flex;
+  gap: 1rem;
+  height: 100vh;
+}
+
+.bpmnContainer {
+  width: 100%;
+}
+
 .container {
   padding: 18px;
   box-sizing: border-box;
 }
 
 .canvasContainer {
-  height: 100vh;
 }
 
 .canvasMenuContainer {
