+++ frontend/packages/process-editor/src/components/Canvas/VersionAlert/VersionAlert.module.css
@@ -1,0 +1,8 @@
+.alert {
+  height: max-content;
+}
+
+.alertText {
+  white-space: pre-line;
+  word-wrap: break-word;
+}
