+++ frontend/packages/process-editor/src/components/Canvas/VersionAlert/index.ts
@@ -1,0 +1,1 @@
+export { VersionAlert } from './VersionAlert';
