+++ frontend/packages/shared/src/utils/featureToggleUtils.ts
@@ -10,7 +10,8 @@ export type SupportedFeatureFlags =
   | 'processEditor'
   | 'configureLayoutSet'
   | 'newAdministration'
-  | 'formTree';
+  | 'formTree'
+  | 'shouldOverrideAppLibCheck';
 
 /*
  * Please add all the features that you want to be toggle on by default here.
